pyproject.toml
[project]
name = "desafio"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "attrs==25.4.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "build==1.3.0",
    "cachetools==6.2.4",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "chromadb==1.4.0",
    "click==8.3.1",
    "coloredlogs==15.0.1",
    "distro==1.9.0",
    "durationpy==0.10",
    "exceptiongroup==1.3.1",
    "fastmcp>=2.14.1",
    "filelock==3.20.1",
    "flatbuffers==25.12.19",
    "fsspec==2025.12.0",
    "google>=3.0.0",
    "google-auth==2.45.0",
    "googleapis-common-protos==1.72.0",
    "grpcio==1.76.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "huggingface-hub==0.36.0",
    "humanfriendly==10.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "importlib-resources==6.5.2",
    "iniconfig==2.3.0",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "jsonschema-specifications==2025.9.1",
    "kubernetes==34.1.0",
    "langchain>=1.2.0",
    "langchain-core>=1.2.5",
    "langchain-google-genai>=4.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mcp[cli]>=1.25.0",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "numpy==2.2.6",
    "oauthlib==3.3.1",
    "onnxruntime==1.23.2",
    "openai>=2.14.0",
    "opentelemetry-api==1.39.1",
    "opentelemetry-exporter-otlp-proto-common==1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.39.1",
    "opentelemetry-proto==1.39.1",
    "opentelemetry-sdk==1.39.1",
    "opentelemetry-semantic-conventions==0.60b1",
    "orjson==3.11.5",
    "overrides==7.7.0",
    "packaging==25.0",
    "pluggy==1.6.0",
    "posthog==5.4.0",
    "protobuf==6.33.2",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.3",
    "pydantic>=2.12.5",
    "pygments==2.19.2",
    "pymupdf==1.26.7",
    "pypdf>=6.5.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "readability-lxml>=0.8.4.1",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rich==14.2.0",
    "rpds-py==0.30.0",
    "rsa==4.9.1",
    "safetensors==0.7.0",
    "scikit-learn==1.7.2",
    "scipy==1.15.3",
    "sentence-transformers==5.2.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tokenizers==0.22.1",
    "tomli==2.3.0",
    "torch==2.9.1",
    "tqdm==4.67.1",
    "transformers==4.57.3",
    "typer==0.21.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.3.0",
    "uvicorn==0.40.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "zipp==3.23.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
